refactor(pulse): extract pure utils to src/lib/utils.ts
Move formatTime, toDatetimeLocal, composeLabel, and PhotonFeature out of
App.tsx so they can be imported in tests.
